+++ src/app/inventory/store/d2-store-factory.service.ts
@@ -1,100 +1,25 @@
-import { copy as angularCopy, IPromise } from 'angular';
+import { copy as angularCopy } from 'angular';
 import {
   DestinyCharacterComponent,
-  DestinyClass,
   DestinyItemComponent,
-  DestinyProgression,
   DestinyStatDefinition
   } from 'bungie-api-ts/destiny2';
 import * as _ from 'underscore';
 import uuidv4 from 'uuid/v4';
 import { bungieNetPath } from '../../dim-ui/bungie-image';
 import { count, sum } from '../../util';
-import { D2InventoryBucket, D2InventoryBuckets } from '../../destiny2/d2-buckets.service';
+import { D2InventoryBuckets } from '../../destiny2/d2-buckets.service';
 import { D2ManifestDefinitions, LazyDefinition } from '../../destiny2/d2-definitions.service';
 import { Loadout } from '../../loadout/loadout.service';
 import { getClass } from './character-utils';
-import { DimItem } from './d2-item-factory.service';
-import { DimStore } from './d2-store-factory.service';
 // tslint:disable-next-line:no-implicit-dependencies
 import vaultBackground from 'app/images/vault-background.png';
 // tslint:disable-next-line:no-implicit-dependencies
 import vaultIcon from 'app/images/vault.png';
 import { showInfoPopup } from '../../shell/info-popup';
 import { t } from 'i18next';
-
-export interface DimCharacterStat {
-  id: number;
-  name: string;
-  description: string;
-  value: number;
-  icon: string;
-  tiers: number[];
-  tierMax: number;
-  tier: number;
-}
-
-export interface DimStore {
-  id: string;
-  name: string;
-  items: DimItem[];
-  isVault: boolean;
-  vault?: DimVault;
-  buckets: { [bucketId: number]: DimItem[] };
-
-  destinyVersion: 1 | 2;
-  icon: string;
-  current: boolean;
-  lastPlayed: Date;
-  background: string;
-  level: number;
-  percentToNextLevel: number;
-  powerLevel: number;
-  stats: DimCharacterStat[];
-  class: 'titan' | 'warlock' | 'hunter';
-  classType: DestinyClass;
-  className: string;
-  gender: string;
-  genderRace: string;
-  progression: null | {
-    progressions: { [key: number]: DestinyProgression };
-  };
-
-  updateCharacterInfo(defs: D2ManifestDefinitions, bStore: DestinyCharacterComponent): IPromise<DimStore[]>;
-  amountOfItem(item: DimItem): number;
-  /**
-   * How much of items like this item can fit in this store? For
-   * stackables, this is in stacks, not individual pieces.
-   */
-  capacityForItem(item: DimItem): number;
-  /**
-   * How many *more* items like this item can fit in this store?
-   * This takes into account stackables, so the answer will be in
-   * terms of individual pieces.
-   */
-  spaceLeftForItem(item: DimItem): number;
-
-  /** Remove an item from this store. Returns whether it actually removed anything. */
-  removeItem(item: DimItem): boolean;
-
-  addItem(item: DimItem): void;
-
-  // Create a loadout from this store's equipped items
-  // TODO: Loadout type
-  loadoutFromCurrentlyEquipped(name: string): Loadout;
-
-  // TODO: implement this for D2
-  factionAlignment();
-}
-
-export interface DimVault extends DimStore {
-  d2VaultCounts: { [bucketId: number]: { count: number; bucket: D2InventoryBucket } };
-  vaultCounts: { [category: string]: number };
-
-  legendaryMarks: number;
-  glimmer: number;
-  silver: number;
-}
+import { D2Store, D2Vault, D2CharacterStat } from '../store-types';
+import { D2Item } from '../item-types';
 
 /**
  * A factory service for producing "stores" (characters or the vault).
@@ -107,7 +32,7 @@ const StoreProto = {
    * Get the total amount of this item in the store, across all stacks,
    * excluding stuff in the postmaster.
    */
-  amountOfItem(this: DimStore, item: DimItem) {
+  amountOfItem(this: D2Store, item: D2Item) {
     return sum(this.items, (i) => {
       return (i.hash === item.hash && !i.location.inPostmaster) ? i.amount : 0;
     });
@@ -117,7 +42,7 @@ const StoreProto = {
    * How much of items like this item can fit in this store? For
    * stackables, this is in stacks, not individual pieces.
    */
-  capacityForItem(this: DimStore, item: DimItem) {
+  capacityForItem(this: D2Store, item: D2Item) {
     if (!item.bucket) {
       console.error("item needs a 'bucket' field", item);
       return 10;
@@ -130,7 +55,7 @@ const StoreProto = {
    * This takes into account stackables, so the answer will be in
    * terms of individual pieces.
    */
-  spaceLeftForItem(this: DimStore, item: DimItem) {
+  spaceLeftForItem(this: D2Store, item: D2Item) {
     if (!item.type) {
       throw new Error("item needs a 'type' field");
     }
@@ -150,7 +75,7 @@ const StoreProto = {
     }
   },
 
-  updateCharacterInfo(this: DimStore, defs: D2ManifestDefinitions, character: DestinyCharacterComponent) {
+  updateCharacterInfo(this: D2Store, defs: D2ManifestDefinitions, character: DestinyCharacterComponent) {
     this.level = character.levelProgression.level; // Maybe?
     this.powerLevel = character.light;
     this.background = bungieNetPath(character.emblemBackgroundPath);
@@ -159,7 +84,7 @@ const StoreProto = {
   },
 
   // Remove an item from this store. Returns whether it actually removed anything.
-  removeItem(this: DimStore, item: DimItem) {
+  removeItem(this: D2Store, item: D2Item) {
     // Completely remove the source item
     const match = (i) => item.index === i.index;
     const sourceIndex = this.items.findIndex(match);
@@ -179,7 +104,7 @@ const StoreProto = {
     return false;
   },
 
-  addItem(this: DimStore, item: DimItem) {
+  addItem(this: D2Store, item: D2Item) {
     this.items.push(item);
     const bucketItems = this.buckets[item.location.id];
     bucketItems.push(item);
@@ -199,8 +124,8 @@ const StoreProto = {
   },
 
   // Create a loadout from this store's equipped items
-  loadoutFromCurrentlyEquipped(this: DimStore, name: string): Loadout {
-    const allItems = (this.items as DimItem[])
+  loadoutFromCurrentlyEquipped(this: D2Store, name: string): Loadout {
+    const allItems = (this.items as D2Item[])
       .filter((item) => item.canBeInLoadout())
       // tslint:disable-next-line:no-unnecessary-callback-wrapper
       .map((item) => angularCopy(item));
@@ -212,12 +137,16 @@ const StoreProto = {
     };
   },
 
-  factionAlignment(this: DimStore) {
-    return null;
+  isDestiny1(this: D2Store) {
+    return false;
+  },
+
+  isDestiny2(this: D2Store) {
+    return true;
   }
 };
 
-export function makeCharacter(defs: D2ManifestDefinitions, character: DestinyCharacterComponent, mostRecentLastPlayed: Date): DimStore {
+export function makeCharacter(defs: D2ManifestDefinitions, character: DestinyCharacterComponent, mostRecentLastPlayed: Date): D2Store {
   const race = defs.Race[character.raceHash];
   const gender = defs.Gender[character.genderHash];
   const classy = defs.Class[character.classHash];
@@ -226,7 +155,7 @@ export function makeCharacter(defs: D2ManifestDefinitions, character: DestinyCha
   const genderName = gender.displayProperties.name;
   const lastPlayed = new Date(character.dateLastPlayed);
 
-  const store: DimStore = Object.assign(Object.create(StoreProto), {
+  const store: D2Store = Object.assign(Object.create(StoreProto), {
     destinyVersion: 2,
     id: character.characterId,
     icon: bungieNetPath(character.emblemPath),
@@ -250,7 +179,7 @@ export function makeCharacter(defs: D2ManifestDefinitions, character: DestinyCha
   return store;
 }
 
-export function makeVault(buckets: D2InventoryBuckets, profileCurrencies: DestinyItemComponent[]): DimVault {
+export function makeVault(buckets: D2InventoryBuckets, profileCurrencies: DestinyItemComponent[]): D2Vault {
   const glimmer = profileCurrencies.find((cur) => cur.itemHash === 3159615086);
   const legendary = profileCurrencies.find((cur) => cur.itemHash === 1022552290);
   const silver = profileCurrencies.find((cur) => cur.itemHash === 3147280338);
@@ -276,20 +205,20 @@ export function makeVault(buckets: D2InventoryBuckets, profileCurrencies: Destin
     silver: currencies.silver,
     isVault: true,
     // Vault has different capacity rules
-    capacityForItem(item: DimItem) {
+    capacityForItem(this: D2Vault, item: D2Item) {
       if (!item.bucket) {
         throw new Error("item needs a 'bucket' field");
       }
       const vaultBucket = buckets.byHash[item.bucket.hash].vaultBucket;
       return vaultBucket ? vaultBucket.capacity : 0;
     },
-    spaceLeftForItem(item: DimItem) {
+    spaceLeftForItem(this: D2Vault, item: D2Item) {
       if (!item.bucket.vaultBucket) {
         return 0;
       }
       const vaultBucket = item.bucket.vaultBucket;
       const usedSpace = item.bucket.vaultBucket
-        ? count((this.items as DimItem[]), (i) => i.bucket.vaultBucket && (i.bucket.vaultBucket.id === vaultBucket.id))
+        ? count((this.items as D2Item[]), (i) => i.bucket.vaultBucket && (i.bucket.vaultBucket.id === vaultBucket.id))
         : 0;
       const openStacks = Math.max(0, this.capacityForItem(item) - usedSpace);
       const maxStackSize = item.maxStackSize || 1;
@@ -301,14 +230,14 @@ export function makeVault(buckets: D2InventoryBuckets, profileCurrencies: Destin
         return (openStacks * maxStackSize) + stackSpace;
       }
     },
-    removeItem(item): DimItem {
+    removeItem(this: D2Vault, item): D2Item {
       const result = StoreProto.removeItem.call(this, item);
       if (item.location.vaultBucket) {
         this.d2VaultCounts[item.location.vaultBucket.id].count--;
       }
       return result;
     },
-    addItem(item: DimItem) {
+    addItem(this: D2Vault, item: D2Item) {
       StoreProto.addItem.call(this, item);
       if (item.location.vaultBucket) {
         this.d2VaultCounts[item.location.vaultBucket.id].count++;
@@ -325,23 +254,20 @@ function getCharacterStatsData(
   stats: {
       [key: number]: number;
   }
-): { [hash: number]: DimCharacterStat } {
+): { [hash: number]: D2CharacterStat } {
   const statWhitelist = [2996146975, 392767087, 1943323491];
-  const ret: { [hash: number]: DimCharacterStat } = {};
+  const ret: { [hash: number]: D2CharacterStat } = {};
 
   // Fill in missing stats
   statWhitelist.forEach((statHash) => {
     const def = statDefs.get(statHash);
     const value = stats[statHash] || 0;
-    const stat: DimCharacterStat = {
+    const stat: D2CharacterStat = {
       id: statHash,
       name: def.displayProperties.name,
       description: def.displayProperties.description,
       value,
-      icon: bungieNetPath(def.displayProperties.icon),
-      tiers: [value],
-      tierMax: 10,
-      tier: 0
+      icon: bungieNetPath(def.displayProperties.icon)
     };
     ret[statHash] = stat;
   });
