+++ src/global.d.ts
@@ -53,6 +53,11 @@ declare module "*/CHANGELOG.md" {
   export default value;
 }
 
+declare module "*.json" {
+  const value: any;
+  export default value;
+}
+
 declare module "*.jpg" {
   const value: string;
   export default value;
